// types/MyInfo/Qualifications.ts
export type Experiences= {
    company: string;
    jobTitle: string;
    from: string;
    to: string;
    comment: string;
}

export type Educations= {
    level: string;
    institute: string;
    majorSpecialization: string;
    year: string;
    gpaScore: string;
    startDate: string;
    endDate: string;
}
export type Skills= {
    skill: string;
    yearsOfExperience: string;
    commentSkills: string;
}
export type Languages= {
    language: string;
    fluency: string;
    competency: string;
    commentLanguage: string;
}
export type Licenses= {
    licenseType: string;
    licenseNumber: string;
    issuedDate: string;
    expiryDate: string;
}